Performace - check if previous user had extension enabled before updating db, need review, maybe it's possible to get old user handle better
git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@104435 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ OnlineStatusBarHooks.php
@@ -22,14 +22,22 @@ class OnlineStatusBarHooks {
 	}
 
 	/**
+	 * Called everytime when user logout
 	 * @param $user User
 	 * @param $inject_html string
 	 * @param $old_name string
 	 * @return bool
 	 */
 	public static function logout( &$user, &$inject_html, $old_name ) {
-		OnlineStatusBar::purge( $old_name );
-		OnlineStatusBar_StatusCheck::deleteStatus( $old_name );
+		global $wgOnlineStatusBarDefaultEnabled;
+		// check if user had enabled this feature before we write to db
+		$olduser = User::newFromName( $old_name );
+		if ( $olduser instanceof User ) {
+			if ( $olduser->getOption( 'OnlineStatusBar_active', $wgOnlineStatusBarDefaultEnabled ) ) {
+				OnlineStatusBar::purge( $old_name );
+				OnlineStatusBar_StatusCheck::deleteStatus( $old_name );
+			}
+		}
 		return true;
 	}
